fixed folder organization + began fixing concurrency issues
