+++ Sources/RoboKit/Socket/MessageModel.swift
